+++ src/lib/components/ContractSelectSection.tsx
@@ -123,7 +123,7 @@ export const ContractSelectSection = observer(
     });
 
     const { refetch } = useQuery(
-      ["query", "instantiateInfo", contractAddress],
+      ["query", "instantiateInfo", endpoint, contractAddress],
       async () =>
         queryInstantiateInfo(endpoint, indexerGraphClient, contractAddress),
       {